+++ sdk/state/update.go
@@ -5,31 +5,25 @@ import (
 	"time"
 
 	"github.com/stellar/experimental-payment-channels/sdk/txbuild"
-	"github.com/stellar/go/keypair"
 	"github.com/stellar/go/xdr"
 )
 
 type PaymentProposal struct {
 	IterationNumber            int64
 	ObservarvationPeriodTime   time.Duration
-	ObservationPeriodLedgerGap int64 // TODO - int64 right?
+	ObservationPeriodLedgerGap int64
 	AmountToInitiator          int64
 	AmountToResponder          int64
 	CloseSignatures            xdr.DecoratedSignature
 	DeclarationSignatures      xdr.DecoratedSignature
 }
 
-// TODO - when to validate the payments?
-func (c *Channel) ValidatePayment(p *PaymentProposal, expectedPayToInitiator int64, expectedPayToResponder int64) bool {
-	return true
-}
-
-// TODO - validate inputs (eg. no negative amounts)
-// initiator always creates the proposals
+// TODO - validate inputs? (eg. no negative amounts)
+// initiator will only call this
 func (c *Channel) NewPaymentProposal(me Participant, other Participant, amountToInitiator int64, amountToResponder int64, startSequence int64, i int64, e int64, o time.Duration, observationPeriodLedgerGap int64, networkPassphrase string) (*PaymentProposal, error) {
 	txC, err := txbuild.Close(txbuild.CloseParams{
 		ObservationPeriodTime:      o,
-		ObservationPeriodLedgerGap: 0,
+		ObservationPeriodLedgerGap: observationPeriodLedgerGap,
 		InitiatorSigner:            me.KP.FromAddress(),
 		ResponderSigner:            other.KP.FromAddress(),
 		InitiatorEscrow:            me.Escrow.FromAddress(),
@@ -42,19 +36,12 @@ func (c *Channel) NewPaymentProposal(me Participant, other Participant, amountTo
 	if err != nil {
 		return nil, err
 	}
-
 	txC, err = txC.Sign(networkPassphrase, initiator.KP)
 	if err != nil {
 		return nil, err
 	}
-
-	// TODO - when should this happen?
-	// c.ProposalStatus = ProposalStatusProposed
-	// TODO - when should the channel balance be updated?
-	// c.Balance += amountToResponder
-	// c.Balance -= amountToInitiator
-
-	// TODO - how to get the transaction signature?
+	c.Balance += amountToResponder
+	c.Balance -= amountToInitiator
 	return &PaymentProposal{
 		IterationNumber:            i,
 		ObservationPeriodTime:      observationPeriodTime,
@@ -67,15 +54,43 @@ func (c *Channel) NewPaymentProposal(me Participant, other Participant, amountTo
 
 // TODO - what do to when you don't have the full other's KP? - Change to FullAddress?
 func (c *Channel) ConfirmPayment(p *PaymentProposal, initiator Participant, responder Participant, isInitiator bool, networkPassphrase string) (*PaymentProposal, error) {
-
 	if !isInitiator {
-		// 1. First confirmation:
-		// TODO - validate P_i
-		// build D_i based off of P_i
-		// build C_i based off of P_i
-		// check that the signatures from P_i are correct
-		// sign C_i and D_i
-		// send something back (P_i with your new signature?)
+		newBalance := c.Balance + p.AmountToResponder - p.AmountToInitiator
+		// TODO - better var names to differentiate from C_i fields?
+		amountToInitiator := int64(0)
+		amountToResponder := int64(0)
+		if newBalance > 0 {
+			amountToResponder = newBalance
+		} else {
+			amountToInitiator = newBalance
+		}
+		txC, err := txbuild.Close(txbuild.CloseParams{
+			ObservationPeriodTime:      p.ObservationPeriodTime,
+			ObservationPeriodLedgerGap: p.ObservationPeriodLedgerGap,
+			InitiatorSigner:            initiator.KP.FromAddress(),
+			ResponderSigner:            responder.KP.FromAddress(),
+			InitiatorEscrow:            initiator.Escrow.FromAddress(),
+			ResponderEscrow:            responder.Escrow.FromAddress(),
+			StartSequence:              p.StartSequence,
+			IterationNumber:            p.IterationNumber,
+			AmountToInitiator:          p.AmountToInitiator,
+			AmountToResponder:          p.AmountToResponder,
+		})
+		if err != nil {
+			return nil, err
+		}
+		if !verifyTxSignatures(txC, p.CloseSignatures) {
+			return nil, errors.New("invalid declaration transaction")
+		}
+		txC, err = txC.AddSignatureDecorated(p.CloseSignatures)
+		if err != nil {
+			return nil, err
+		}
+		// TODO - why is signing here bad?
+		txC, err = txC.Sign(networkPassphrase, me.KP)
+		if err != nil {
+			return nil, err
+		}
 
 		txD, err := txbuild.Declaration(txbuild.DeclarationParams{
 			InitiatorEscrow:         initiator.Escrow.FromAddress(),
@@ -86,38 +101,12 @@ func (c *Channel) ConfirmPayment(p *PaymentProposal, initiator Participant, resp
 		if err != nil {
 			return nil, err
 		}
-
-		if !verifyTxSignatures(txD, p.CloseSignatures) {
-			return nil, errors.New("invalid declaration transaction")
-		}
-
-		// TODO - add the given signature to your new created transaction
-
 		txD, err := txD.Sign(networkPassphrase, me.KP)
 		if err != nil {
 			return nil, err
 		}
 
-		txC, err := txbuild.Close(txbuild.CloseParams{
-			ObservationPeriodTime:      p.ObservationPeriodTime,
-			ObservationPeriodLedgerGap: p.ObservationPeriodLedgerGap,
-			InitiatorSigner:            initiator.KP.FromAddress(),
-			ResponderSigner:            responder.KP.FromAddress(),
-			InitiatorEscrow:            initiator.Escrow.FromAddress(),
-			ResponderEscrow:            responder.Escrow.FromAddress(),
-			StartSequence:              p.StartSequence, // TODO - should this be on P?
-			IterationNumber:            p.IterationNumber,
-			AmountToInitiator:          p.AmountToInitiator,
-			AmountToResponder:          p.AmountToResponder,
-		})
-		if err != nil {
-			return nil, err
-		}
-		// TODO - why is signing here bad?
-		txC, err = txC.Sign(networkPassphrase, me.KP)
-		if err != nil {
-			return nil, err
-		}
+		c.Balance = newBalance
 		p.CloseSignatures = txC.Signatures()
 		p.DeclarationSignatures = txD.Signatures()
 		return p
@@ -141,16 +130,3 @@ func (c *Channel) ConfirmPayment(p *PaymentProposal, initiator Participant, resp
 	p.DeclarationSignatures = txD.Signatures()
 	return p
 }
-
-// Common data both participants will have during the test.
-type Participant struct {
-	Name                 string
-	KP                   *keypair.Full
-	Escrow               *keypair.Full
-	EscrowSequenceNumber int64
-	Contribution         int64
-}
-
-// TODO
-// - ProposalStatus is best?
-// - message validation
